ml-service/requirements.txt
fastapi
uvicorn
pydantic
chromadb
numpy
opencv-python
pillow
python-multipart
aiofiles
tqdm
requests
torch
torchvision
open-clip-torch
ftfy
regex
